Refactoring to reduce duplicate code in ValCommandTests.
