+++ Tests/ValCommandTests/Utils/String+Log.swift
@@ -1,0 +1,7 @@
+import ValCommand
+
+extension String: Log {
+
+  public var hasANSIColorSupport: Bool { false }
+
+}
