+++ Tests/ValCommandTests/ValCommandTests.swift
@@ -4,20 +4,6 @@ import XCTest
 
 final class ValCommandTests: XCTestCase {
 
-  /// The result of a compiler invokation.
-  private struct CompilationResult {
-
-    /// The exit status of the compiler.
-    let status: ExitCode
-
-    /// The URL of the output file.
-    let output: URL
-
-    /// The contents of the standard error.
-    let stderr: String
-
-  }
-
   func testNoInput() throws {
     XCTAssertThrowsError(try ValCommand.parse([]))
   }
@@ -25,32 +11,25 @@ final class ValCommandTests: XCTestCase {
   func testRawAST() throws {
     let input = try url(forSourceNamed: "Success")
     let result = try compile(input, with: ["--emit", "raw-ast"])
-    XCTAssert(result.status.isSuccess)
-    XCTAssert(FileManager.default.fileExists(atPath: result.output.relativePath))
-    XCTAssert(result.stderr.isEmpty)
+    result.assertSuccess()
   }
 
   func testRawIR() throws {
     let input = try url(forSourceNamed: "Success")
     let result = try compile(input, with: ["--emit", "raw-ir"])
-    XCTAssert(result.status.isSuccess)
-    XCTAssert(result.stderr.isEmpty)
-    XCTAssert(FileManager.default.fileExists(atPath: result.output.relativePath))
+    result.assertSuccess()
   }
 
   func testIR() throws {
     let input = try url(forSourceNamed: "Success")
     let result = try compile(input, with: ["--emit", "ir"])
-    XCTAssert(result.status.isSuccess)
-    XCTAssert(result.stderr.isEmpty)
-    XCTAssert(FileManager.default.fileExists(atPath: result.output.relativePath))
+    result.assertSuccess()
   }
 
   func testCPP() throws {
     let input = try url(forSourceNamed: "Success")
     let result = try compile(input, with: ["--emit", "cpp"])
-    XCTAssert(result.status.isSuccess)
-    XCTAssert(result.stderr.isEmpty)
+    result.assertSuccess(checkOutput: false)
 
     let baseURL = result.output.deletingPathExtension()
     XCTAssert(
@@ -62,8 +41,7 @@ final class ValCommandTests: XCTestCase {
   func testTypeCheckSuccess() throws {
     let input = try url(forSourceNamed: "Success")
     let result = try compile(input, with: ["--typecheck"])
-    XCTAssert(result.status.isSuccess)
-    XCTAssert(result.stderr.isEmpty)
+    result.assertSuccess(checkOutput: false)
   }
 
   func testTypeCheckFailure() throws {
@@ -91,34 +69,5 @@ final class ValCommandTests: XCTestCase {
       "No inputs",
       file: file, line: line)
   }
-
-  /// Compiles `input` with the given arguments and returns the compiler's exit status, the URL of
-  /// the output file, and the contents of the standard error.
-  private func compile(
-    _ input: URL,
-    with arguments: [String]
-  ) throws -> CompilationResult {
-    // Create a temporary directory to write the output file.
-    let outputDirectory = try FileManager.default.url(
-      for: .itemReplacementDirectory,
-      in: .userDomainMask,
-      appropriateFor: input,
-      create: true)
-    let output = outputDirectory.appendingPathComponent("a.out")
-
-    // Parse the command line's arguments.
-    let cli = try ValCommand.parse(arguments + ["-o", output.relativePath, input.relativePath])
-
-    // Execute the command.
-    var stderr = ""
-    let status = try cli.execute(loggingTo: &stderr)
-    return CompilationResult(status: status, output: output, stderr: stderr)
-  }
-
-}
-
-extension String: Log {
-
-  public var hasANSIColorSupport: Bool { false }
 
 }
